Reformatted Calculator Logic & Many Unit Test Changes
So much is just working, WOOOOOOOO. Changed input types of most calculator logic (expect 1 but thats after lunch) to strings from floats and added checks to ensure that we get something that is readable by the code (aka a number). Unit Tests were updated to reflect the new input types.